backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
httpx>=0.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
